refactor: move PcapProcessor into pcapprocessor/processor.py, fix dynamic import
Replace the non-standard importlib.util dynamic loading pattern with a clean
standard import. The PcapProcessor class is now defined in pcapprocessor/processor.py,
eliminating the naming collision between the root script and the package directory.

- Create pcapprocessor/processor.py with PcapProcessor class
- Update root pcapprocessor.py to import from package (CLI entry point preserved)
- Update pcapprocessor/__init__.py to use standard import instead of importlib.util

All 20 tests pass. Imports work from both package and module paths.
